configs/requirements.txt
pyyaml
numpy
pandas
tqdm
scikit-learn
sentence-transformers
faiss-cpu
torch
transformers
accelerate
peft
matplotlib
py7zr
mwparserfromhell

